+++ app/__init__.py
@@ -1,11 +1,9 @@
-import os
-from flask import Flask, render_template, request
-from dotenv import load_dotenv
+from flask import Flask
+from views import views
 
-load_dotenv()
 app = Flask(__name__)
 
+app.register_blueprint(views, url_prefix="/")
 
-@app.route('/')
-def index():
-    return render_template('index.html', title="MLH Fellow", url=os.getenv("URL"))
+if __name__ == "__main__":
+    app.run(debug=True)
