--- templates/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<head>
-    <meta charset="UTF-8">
-    <title>Home</title>
-</head>
-<body>
-    <div><h1>Hi, i'm {{name}}</h1></div>
-    <p>{{about}}</p>
-
-</body>
-</html>
